Swagger Doc Buildout
Shifted swagger docs out of js files and into yml

Added more docs
