+++ api-server/routes/announcements/NewAnnouncementRequest.yml
@@ -1,0 +1,13 @@
+definitions:
+  NewAnnouncementRequest:
+    type: object
+    required:
+      - announcement
+    properties:
+      announcement:
+        type: object
+        required:
+          - message
+        properties:
+          message:
+            type: string
