+++ api-server/routes/events/handlers/getCurrentVerifiedEvents.js
@@ -1,38 +1,3 @@
-/**
- * @swagger
- *
- * /events/current/verified:
- *  get:
- *    description: Returns a list of all verified events. Currents events are events that are upcoming or less than 24 hours in the past. This is an open route for public consumption.
- *    produces: application/json
- *    parameters:
- *      - name: embed
- *        description: |+
- *          Specifies related entities to retrieve as embedded children of the event. For example, events have venues.
- *
- *          Allowed Values -- venue
- *        in: query
- *        required: false
- *        type: array
- *        items:
- *          type: string
- *        enum:
- *          - venue
- *    responses:
- *      200:
- *        description: Success!
- *        schema:
- *          $ref: '#definitions/EventsResponse'
- *      501:
- *        description: There was an error processing the request.
- *        schema:
- *          $ref: '#definitions/EventsResponse'
- *      422:
- *        description: A parameter supplied was not allowed or understood.
- *        schema:
- *          $ref: '#definitions/EventsResponse'
- */
-
 const  { literal } = require('sequelize')
 const express = require('express')
