+++ api-server/routes/events/handlers/verifyEventById.yml
@@ -1,0 +1,25 @@
+/events/verify/{id}:
+  put:
+    description: Sets the specified event to verified
+    produces: application/json
+    security:
+      - jwt:
+    parameters:
+      - name: id
+        description: id of event to verify
+        in: path
+        required: true
+        type: string
+    responses:
+      200:
+        description: Success!
+        schema:
+          $ref: '#definitions/EventsResponse'
+      501:
+        description: There was an error processing the request.
+        schema:
+          $ref: '#definitions/EventsResponse'
+      422:
+        description: A parameter supplied was not allowed or understood.
+        schema:
+          $ref: '#definitions/EventsResponse'
