+++ api-server/routes/icsFile/helpers/icsFileSettings.js
@@ -1,0 +1,5 @@
+module.exports = {
+  calscale: 'GREGORIAN',
+  prodid: 'InfiniteIndustries/ics',
+  timezone: 'America/New_York'   // need to figure out a more graceful way to deal with timezones
+}
