+++ api-server/routes/users/userResponse.yml
@@ -1,0 +1,13 @@
+definitions:
+  UserResponse:
+    type: object
+    properties:
+      status:
+        type: string
+        enum: ["success", "failure"]
+      error_message:
+        type: string
+      users:
+        type: array
+        items:
+          $ref: '#definitions/User'
